linux: disable PCI to PCI bridge
The presence of the PCI bridge introduces a regression when booting
OVMF:

  PciHostBridgeDxe: IntersectMemoryDescriptor: desc [0, A0000) type 2 cap
  800000000002600F conflicts with aperture [0, D0021000) cap 1

The root cause is that the bridge does not correctly populate the memory
base/limit and prefetchable memory base/limit PCI bridge configuration
registers.

For the short term, since this bridge is only needed for the PCI hotplug
feature that is not enabled yet, remove the creation of the bridge from
the startup code.  Fixing the bridge will be done in another commit to
keep this change small and easy to backport.

BUG=b:199442120
BUG=b:185084350
TEST=crosvm run --bios OVMF.fd

Change-Id: I91b2cc6628e320ac2403bd8b74d18ef5a26d7c0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3158942
Reviewed-by: Xiong  Zhang <xiong.y.zhang@intel.corp-partner.google.com>
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>