rutabaga_gfx: add better error reporting
Both RutabagaError::SpecViolation and RutabagaError::Unsupported
are used as catch-alls for errors that don't fit any one
particular place.

With this change, RutabagaError::Unsupported means that a particular
function is not implemented.

RutabagaError::SpecViolation(..) now provides free-form debug
reporting for one-off errors.  However, it is recommended to create a
new RutabagaError variant if two instances of the same spec violation
are noticed.  Many new variants are added in CL in observance of this
recommendation.

In addition, we may create a CrossDomainError and RutabagaGrallocError
one day.  Currently, it doesn't make sense to do so at the moment.
But prefix the new variants accordingly if the need arises.

BUG=b:173630595
TEST=compile and run

Change-Id: Ie610c77445f4406b40bb40f00bcabb8e65c8e02d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3203231
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Jason Macnak <natsu@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>