devices: re-add set_access_id to bus
We need the access id for some future changes that will be posted soon
for a userspace irqchip.

Each vcpu will have a clone of the Bus itself, not the Arc<Bus>, so that
they can each have their own access id.

This also adds a call to set_access_id to run_vcpu.

BUG=b:160610704
TEST=ran local presubmit, and ran test image

Change-Id: I4afb0ef9c234f25150ce2f89d6fb228ad5b88745
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3261627
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Colin Downs-Razouk <colindr@google.com>