Revert "x86_64: remove acpi=noirq kernel option"
Basic functionality works with a purely ACPI-based boot (no MPTable),
but something is still missing in the ACPI tables for describing the
full interrupt routing configuration. The guest kernel complains:

  virtio-pci 0000:00:01.0: can't derive routing for PCI INT A
  virtio-pci 0000:00:01.0: PCI INT A: no GSI

Restore the acpi=noirq option for now until we fill out the ACPI IRQ
descriptors correctly; this was a cleanup only and not required to fix a
bug.

This reverts commit cb464a7c083ac71e1ec607872f924b7310c3883a.

BUG=None
TEST=Boot Crostini; no interrupt routing failures in dmesg

Change-Id: Ib73c5749b48134124b3ebcb472706f087e8a85a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3315226
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>