vmm_vhost: Remove SlaveReqHandler type parameter in SlaveReqHelper
The type parameter `S: SlaveReqHandler` for `SlaveReqHelper` was
introduced to check if it's vvu or not.
Howver, SlaveReqHelper shouldn't take a type parameter of
`SlaveReqHandler` because the helper will be used in the VVU proxy
device which doesn't  have its own req_handler.

So, instead of taking the type parameter, we should have a new bool
field which indicates if it's VVU or not.

BUG=b:194136484
TEST=cargo test --all-features in vmm_vhost
TEST=kokoro

Change-Id: I35f20ef7c42cd74678ab5d87b68bbb58586d2fa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3320122
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>